@@ -1,41 +1,40 @@
 {
-    "name": "smart-second-brain",
-    "version": "1.1.0",
-    "main": "main.js",
-    "devDependencies": {
-        "@biomejs/biome": "1.2.2",
-        "@rollup/plugin-json": "^6.1.0",
-        "@rollup/plugin-typescript": "^11.1.6",
-        "@sveltejs/vite-plugin-svelte": "^2.4.6",
-        "@tsconfig/svelte": "^4.0.1",
-        "@typescript-eslint/eslint-plugin": "^6.19.0",
-        "@typescript-eslint/parser": "^6.19.0",
-        "autoprefixer": "^10.4.16",
-        "builtin-modules": "^3.3.0",
-        "bun-types": "latest",
-        "eslint": "^8.56.0",
-        "obsidian": "^1.4.11",
-        "postcss": "^8.4.30",
-        "svelte": "^4.2.1",
-        "svelte-check": "^3.5.2",
-        "tailwindcss": "^3.3.3",
-        "tslib": "^2.6.2",
-        "typescript": "^5.2.2",
-        "vite": "^4.4.9"
-    },
-    "private": true,
-    "scripts": {
-        "dev": "vite build --mode development --watch",
-        "build": "vite build --mode production",
-        "check": "svelte-check --tsconfig ./tsconfig.json",
-        "format": "biome format --write ./src",
-        "lint": "biome lint --apply-unsafe ./src"
-    },
-    "type": "module",
-    "dependencies": {
-        "i18next": "^23.8.2",
-        "monkey-around": "^2.3.0",
-        "nanoid": "^5.0.4",
-        "svelte-i18n": "^4.0.0"
-    }
+  "name": "smart-second-brain",
+  "version": "0.1.0",
+  "main": "main.js",
+  "devDependencies": {
+    "@biomejs/biome": "^1.9.4",
+    "@sveltejs/vite-plugin-svelte": "^5.1.1",
+    "@tsconfig/svelte": "^5.0.5",
+    "autoprefixer": "^10.4.21",
+    "obsidian": "^1.10.0",
+    "postcss": "^8.5.6",
+    "svelte": "^5.41.0",
+    "svelte-check": "^4.3.3",
+    "tailwindcss": "^3.4.18",
+    "tslib": "^2.8.1",
+    "typescript": "^5.9.3",
+    "vite": "^6.4.0"
+  },
+  "private": true,
+  "scripts": {
+    "dev": "vite build --mode development --watch",
+    "build": "vite build --mode production",
+    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "format": "biome format --write ./src",
+    "lint": "biome lint --apply-unsafe ./src"
+  },
+  "type": "module",
+  "dependencies": {
+    "@selemondev/svgl-svelte": "^0.0.3",
+    "@tanstack/svelte-query": "https://pkg.pr.new/@tanstack/svelte-query@746b9c5d83e901deb3a5ca7012a4b4e2b617c5e3",
+    "bits-ui": "^2.11.8",
+    "builtin-modules": "^5.0.0",
+    "dexie": "^4.2.1",
+    "i18next": "^23.16.8",
+    "monkey-around": "^3.0.0",
+    "papa-ts": "link:papa-ts",
+    "svelte-i18n": "^4.0.1",
+    "uuid": "^11.1.0"
+  }
 }
