@@ -1,94 +1,75 @@
-<script lang="ts">
-    import { setIcon } from 'obsidian';
-    import { plugin, data, chatHistory, papaState, papaIndexingProgress, errorState, papaIndexingTimeLeft, isChatInSidebar } from '../../store';
-    import { Prompts, type Language, Languages } from 'papa-ts';
-    import ProgressBar from '../base/ProgressBar.svelte';
-    import { t } from 'svelte-i18n';
-    import DropdownComponent from '../base/Dropdown.svelte';
-    import ToggleComponent from '../base/Toggle.svelte';
-    import PullOllamaModel from '../Onboarding/PullOllamaModel.svelte';
-    import LoadingAnimation from '../base/LoadingAnimation.svelte';
-    import { ConfirmModal } from '../Settings/ConfirmModal';
-    import { get } from 'svelte/store';
-    import Ollama from '../Settings/Ollama.svelte';
+<!-- <script lang="ts">
+import { run } from "svelte/legacy";
+import {
+	plugin,
+	chatHistory,
+	papaState,
+	papaIndexingProgress,
+	errorState,
+	papaIndexingTimeLeft,
+	isChatInSidebar,
+} from "../../store";
+import { Prompts, type Language, Languages } from "papa-ts";
+import ProgressBar from "../base/ProgressBar.svelte";
+import { t } from "svelte-i18n";
+import DropdownComponent from "../base/Dropdown.svelte";
+import ToggleComponent from "../base/Toggle.svelte";
+import LoadingAnimation from "../base/LoadingAnimation.svelte";
+import { icon } from "../../utils/icons";
+import { getData } from "../../stores/dataStore.svelte";
 
-    const icon = (node: HTMLElement, iconId: string) => {
-        setIcon(node, iconId);
-    };
+const { isQuickSettingsOpen } = plugin.pluginData;
 
-    let isOpen = $data.isQuickSettingsOpen;
+function toggleDrawer() {
+	isOpen = !isOpen;
+	$plugin.saveSettings({ isQuickSettingsOpen: isOpen });
+}
 
-    function toggleDrawer() {
-        isOpen = !isOpen;
-        $data.isQuickSettingsOpen = isOpen;
-        $plugin.saveSettings();
-    }
+let similarityThreshold = $state(Math.round($data.embedModelConfigs[$data.selEmbedModel].similarityThreshold * 100));
+run(() => {
+	similarityThreshold = Math.min(Math.max(similarityThreshold, 0), 100);
+});
+function setSimilarityThreshold() {
+	$data.embedModelConfigs[$data.selEmbedModel].similarityThreshold = similarityThreshold / 100;
+	$plugin.s2b.configure({ embedModelConfig: { similarityThreshold: similarityThreshold / 100 } });
+	$plugin.saveSettings();
+}
 
-    let similarityThreshold = Math.round(
-        ($data.isIncognitoMode ? $data.ollamaEmbedModel.similarityThreshold : $data.openAIEmbedModel.similarityThreshold) * 100
-    );
-    $: similarityThreshold = Math.min(Math.max(similarityThreshold, 0), 100);
-    function setSimilarityThreshold() {
-        if ($data.isIncognitoMode) {
-            $data.ollamaEmbedModel.similarityThreshold = similarityThreshold / 100;
-        } else {
-            $data.openAIEmbedModel.similarityThreshold = similarityThreshold / 100;
-        }
-        $plugin.s2b.setSimilarityThreshold(similarityThreshold / 100);
-        $plugin.saveSettings();
-    }
+let temperature = $state(Math.round($data.genModelConfigs[$data.selGenModel].temperature * 100));
+run(() => {
+	temperature = Math.min(Math.max(temperature, 0), 100);
+});
+function setTemperature() {
+	$data.genModelConfigs[$data.selGenModel].temperature = temperature / 100;
+	$plugin.s2b.configure({ genModelConfig: { temperature: temperature / 100 } });
+	$plugin.saveSettings();
+}
 
-    let temperature = Math.round($data.isIncognitoMode ? $data.ollamaGenModel.temperature : $data.openAIGenModel.temperature) * 100;
-    $: temperature = Math.min(Math.max(temperature, 0), 100);
-    function setTemperature() {
-        if ($data.isIncognitoMode) {
-            $data.ollamaGenModel.temperature = temperature / 100;
-            $plugin.s2b.setGenModel($data.ollamaGenModel);
-        } else {
-            $data.openAIGenModel.temperature = temperature / 100;
-            $plugin.s2b.setGenModel($data.openAIGenModel);
-        }
-        $plugin.saveSettings();
-    }
+const languages: { display: Language; value: Language }[] = Object.values(Languages).map((language: Language) => ({
+	display: language,
+	value: language,
+}));
 
-    const languages: { display: Language; value: Language }[] = Object.values(Languages).map((language: Language) => ({ display: language, value: language }));
+const setAssistantLanguage = (selected: Language) => {
+	if ($chatHistory.length === 1) chatHistory.reset();
+	$plugin.saveSettings({
+		assistantLanguage: selected,
+		initialAssistantMessageContent: Prompts[selected].initialAssistantMessage,
+	});
+};
+function setChatViewDensity() {
+	$plugin.saveSettings({ isChatComfy: !$data.isChatComfy });
+}
 
-    const setAssistantLanguage = (selected: Language) => {
-        $data.assistantLanguage = selected;
-        $data.initialAssistantMessageContent = Prompts[selected].initialAssistantMessage;
-        if ($chatHistory.length === 1) chatHistory.reset();
-
-        $plugin.saveSettings();
-    };
-    function setChatViewDensity() {
-        $data.isChatComfy = !$data.isChatComfy;
-        $plugin.saveSettings();
-    }
-
-    function initSecondBrain() {
-        $data.isIncognitoMode
-            ? $plugin.s2b.init()
-            : new ConfirmModal(
-                  get(plugin).app,
-                  $t('init_third_party_modal.title'),
-                  $t('init_third_party_modal.description'),
-                  (result) => {
-                      if (result === 'Yes') $plugin.s2b.init();
-                  },
-                  'hideIncognitoWarning'
-              ).activate();
-    }
-    function formatTime(secondsInput: number) {
-        const minutes = Math.floor(secondsInput / 60); // Calculate minutes
-        const seconds = secondsInput % 60; // Remaining seconds
-        return minutes > 0
-            ? $t('quick_settings.time_left_minutes', { values: { minutes, seconds } })
-            : $t('quick_settings.time_left_seconds', { values: { seconds } });
-    }
+function formatTime(secondsInput: number) {
+	const minutes = Math.floor(secondsInput / 60); // Calculate minutes
+	const seconds = secondsInput % 60; // Remaining seconds
+	return minutes > 0
+		? $t("quick_settings.time_left_minutes", { values: { minutes, seconds } })
+		: $t("quick_settings.time_left_seconds", { values: { seconds } });
+}
 </script>
 
-<!-- svelte-ignore a11y-click-events-have-key-events -->
-<!-- svelte-ignore a11y-no-static-element-interactions -->
 <div
     class={`relative ${isOpen ? 'h-[33%] min-h-[33%]' : 'h-[--icon-m] min-h-[--icon-m]'} flex justify-center overflow-hidden transition-all duration-300 ease-in-out`}
 >
@@ -98,10 +79,10 @@
                 <h3 class="text-center text-primary">{$t('quick_settings.initialize')}</h3>
                 <button
                     aria-label={$t('quick_settings.initialize')}
-                    on:click={() => $plugin.s2b.init()}
+                    onclick={() => $plugin.s2b.init()}
                     class="h-8 rounded-l-md px-4 py-2 transition duration-300 ease-in-out hover:bg-[--text-accent-hover]"
                     use:icon={'power'}
-                />
+                ></button>
             {:else if $papaState === 'loading'}
                 <LoadingAnimation />
             {:else if $papaState === 'indexing' || $papaState === 'indexing-pause'}
@@ -116,69 +97,67 @@
                     {#if $papaState === 'indexing'}
                         <div
                             aria-label={$t('quick_settings.pause_indexing')}
-                            on:click={() => ($papaState = 'indexing-pause')}
+                            onclick={() => ($papaState = 'indexing-pause')}
                             class="h-8 py-2 pl-3 pr-2 transition duration-300 ease-in-out hover:text-[var(--text-accent-hover)]"
                             use:icon={'pause'}
-                        />
+                        ></div>
                     {:else if $papaState === 'indexing-pause'}
                         <div
                             aria-label={$t('quick_settings.resume_indexing')}
-                            on:click={() => $plugin.s2b.init()}
+                            onclick={() => $plugin.s2b.init()}
                             class="h-8 py-2 pl-3 pr-2 transition duration-300 ease-in-out hover:text-[var(--text-accent-hover)]"
                             use:icon={'play'}
-                        />
+                        ></div>
                     {/if}
                     <div
                         aria-label={$t('quick_settings.cancel_indexing')}
-                        on:click={() => $plugin.s2b.cancelIndexing()}
+                        onclick={() => $plugin.s2b.cancelIndexing()}
                         class="h-8 py-2 pl-2 pr-3 transition duration-300 ease-in-out hover:text-[var(--text-accent-hover)]"
                         use:icon={'stop-circle'}
-                    />
+                    ></div>
                 </div>
             {:else if $papaState === 'error'}
                 {#if $errorState === 'ollama-gen-model-not-installed'}
-                    <h3 class="text-center text-primary">{$t('quick_settings.error.install_model', { values: { model: $data.ollamaGenModel.model } })}</h3>
-                    <PullOllamaModel pullModel={$data.ollamaGenModel.model} onSuccessfulPull={() => ($papaState = 'settings-change')} />
+                    <h3 class="text-center text-primary">{$t('quick_settings.error.install_model', { values: { model: $data.selGenModel } })}</h3>
                 {:else if $errorState === 'ollama-embed-model-not-installed'}
-                    <h3 class="text-center text-primary">{$t('quick_settings.error.install_model', { values: { model: $data.ollamaEmbedModel.model } })}</h3>
-                    <PullOllamaModel pullModel={$data.ollamaEmbedModel.model} onSuccessfulPull={() => ($papaState = 'settings-change')} />
+                    <h3 class="text-center text-primary">{$t('quick_settings.error.install_model', { values: { model: $data.selEmbedModel } })}</h3>
                 {:else if $errorState === 'failed-indexing'}
                     <h3 class="text-center">{$t('notice.failed_indexing')}</h3>
                     <button
                         aria-label={$t('quick_settings.retry_indexing')}
-                        on:click={() => $plugin.s2b.init()}
+                        onclick={() => $plugin.s2b.init()}
                         class="h-8 rounded-l-md px-4 py-2 transition duration-300 ease-in-out hover:bg-[--text-accent-hover]"
                         use:icon={'refresh-cw'}
-                    />
+                    ></button>
                 {:else}
                     <h3 class="text-center text-primary">{$t('quick_settings.error.other')}</h3>
                     <button
                         aria-label={$t('quick_settings.retry_initialization')}
-                        on:click={() => $plugin.s2b.init()}
+                        onclick={() => $plugin.s2b.init()}
                         class="h-8 rounded-l-md px-4 py-2 transition duration-300 ease-in-out hover:bg-[--text-accent-hover]"
                         use:icon={'refresh-cw'}
-                    />
+                    ></button>
                 {/if}
             {:else if $papaState === 'mode-change'}
-                <h3 class="text-center text-primary">{$t('quick_settings.mode_changed')}{$data.isIncognitoMode ? 'Ollama' : 'OpenAI'}.</h3>
+                <h3 class="text-center text-primary">{$t('quick_settings.mode_changed')}{$data.selGenProvider}.</h3>
                 <button
                     aria-label={$t('quick_settings.reinitialize')}
-                    on:click={() => initSecondBrain()}
+                    onclick={() => $plugin.s2b.init()}
                     class="h-8 rounded-l-md px-4 py-2 transition duration-300 ease-in-out hover:bg-[--text-accent-hover]"
                     use:icon={'refresh-cw'}
-                />
+                ></button>
             {:else if $papaState === 'settings-change'}
                 <h3 class="text-center text-primary">{$t('quick_settings.settings_changed')}</h3>
                 <button
                     aria-label={$t('quick_settings.reinitialize')}
-                    on:click={() => $plugin.s2b.init()}
+                    onclick={() => $plugin.s2b.init()}
                     class="h-8 rounded-l-md px-4 py-2 transition duration-300 ease-in-out hover:bg-[--text-accent-hover]"
                     use:icon={'refresh-cw'}
-                />
+                ></button>
             {:else}
-                <h2 class="mb-0 text-primary">{$data.isIncognitoMode ? 'Ollama' : 'OpenAI'}</h2>
+                <h2 class="mb-0 text-primary">{$data.selGenProvider}</h2>
                 <p class="mt-1">
-                    {$t('quick_settings.chat_via', { values: { model: $data.isIncognitoMode ? $data.ollamaGenModel.model : $data.openAIGenModel.model } })}
+                    {$t('quick_settings.chat_via', { values: { model: $data.selGenModel } })}
                 </p>
                 <div class="w-full max-w-[300px]">
                     <div class="flex h-8 items-center justify-between">
@@ -191,26 +170,26 @@
                     </div>
                     <div class="flex h-8 items-center justify-between">
                         {$t('quick_settings.creativity')}
-                        <div use:icon={'help'} aria-label={$t('tooltip.creativity')} class="ml-1 mr-auto h-[18px]" />
+                        <div use:icon={'help'} aria-label={$t('tooltip.creativity')} class="ml-1 mr-auto h-[18px]"></div>
                         <div class="flex items-center">
                             <output>{temperature}%</output>
-                            <input class="slider" type="range" bind:value={temperature} on:blur={setTemperature} min="0" max="100" />
+                            <input class="slider" type="range" bind:value={temperature} onblur={setTemperature} min="0" max="100" />
                         </div>
                     </div>
                     {#if $data.isUsingRag}
                         <div class="flex h-8 items-center justify-between">
                             {$t('quick_settings.similarity_threshold')}
-                            <div use:icon={'help'} aria-label={$t('tooltip.similarity')} class="ml-1 mr-auto h-[18px]" />
+                            <div use:icon={'help'} aria-label={$t('tooltip.similarity')} class="ml-1 mr-auto h-[18px]"></div>
                             <div class="flex items-center">
                                 <output>{similarityThreshold}%</output>
-                                <input class="slider" type="range" bind:value={similarityThreshold} on:blur={setSimilarityThreshold} min="0" max="100" />
+                                <input class="slider" type="range" bind:value={similarityThreshold} onblur={setSimilarityThreshold} min="0" max="100" />
                             </div>
                         </div>
                     {/if}
                 </div>
             {/if}
         {/if}
-        <div style="height: 1rem" />
+        <div style="height: 1rem"></div>
         <div class="absolute bottom-0 z-10 flex w-full justify-center">
             <div
                 aria-label={isOpen ? $t('quick_settings.close') : $t('quick_settings.open')}
@@ -218,8 +197,8 @@
                     isOpen ? 'rotate-180 transform' : 'rotate-0 transform'
                 }`}
                 use:icon={'chevron-down'}
-                on:click={toggleDrawer}
-            />
+                onclick={toggleDrawer}
+            ></div>
         </div>
     </div>
-</div>
+</div> -->
