@@ -1,10 +1,16 @@
 <script lang="ts">
-    import { isChatInSidebar } from '../../store';
-    import { onMount } from 'svelte';
-    import { tweened } from 'svelte/motion';
-    import { cubicOut } from 'svelte/easing';
+    import { run } from "svelte/legacy";
 
-    export let progress = 0; // Progress value from 0 to 100
+    import { onMount } from "svelte";
+    import { tweened } from "svelte/motion";
+    import { cubicOut } from "svelte/easing";
+    import { chatLayout } from "../../stores/state.svelte";
+
+    interface Props {
+        progress?: number; // Progress value from 0 to 100
+    }
+
+    let { progress = 0 }: Props = $props();
     const radius = 45; // Radius of the circle
     const circumference = 2 * Math.PI * radius; // Circumference of the circle
     const progressValue = tweened(0, {
@@ -13,10 +19,14 @@
     });
 
     // Watch for changes in progress and update the progressValue accordingly
-    $: progress, progressValue.set(progress);
+    run(() => {
+        (progress, progressValue.set(progress));
+    });
 
     // Calculate the stroke dashoffset based on the progress
-    $: strokeDashoffset = circumference - (progress / 100) * circumference;
+    let strokeDashoffset = $derived(
+        circumference - (progress / 100) * circumference,
+    );
 
     onMount(() => {
         progressValue.set(progress);
@@ -25,7 +35,9 @@
 
 <svg class="h-6 w-6 -rotate-90" viewBox="0 0 100 100">
     <circle
-        class="stroke-current {$isChatInSidebar ? 'text-[--background-secondary-alt]' : 'text-[--background-primary-alt]'}"
+        class="stroke-current {chatLayout.isSidebar
+            ? 'text-[--background-secondary-alt]'
+            : 'text-[--background-primary-alt]'}"
         fill="none"
         stroke-width="10"
         cx="50"
