@@ -1,10 +1,14 @@
 <script lang="ts">
-    export let isEnabled: boolean = false;
-    export let changeFunc: () => void;
+interface Props {
+	isToggled: boolean;
+	changeFunc: () => void;
+}
+
+let { isToggled = false, changeFunc }: Props = $props();
 </script>
 
-<!-- svelte-ignore a11y-click-events-have-key-events -->
-<!-- svelte-ignore a11y-no-static-element-interactions -->
-<div class="checkbox-container" class:is-enabled={isEnabled} on:click={changeFunc}>
+<!-- svelte-ignore a11y_click_events_have_key_events -->
+<!-- svelte-ignore a11y_no_static_element_interactions -->
+<div class="checkbox-container" class:is-enabled={isToggled} onclick={changeFunc}>
     <input type="checkbox" tabindex="0" />
 </div>
