@@ -1,9 +1,5 @@
-<script>
-    import { data } from '../../store';
-</script>
-
 <svg
-    class="z-[-1] h-full {$data.isUsingRag ? 'fill-[--color-accent] stroke-[--color-accent]' : 'fill-[--text-normal] stroke-[--text-normal]'}"
+    class="z-[-1] h-full {true ? 'fill-[--color-accent] stroke-[--color-accent]' : 'fill-[--text-normal] stroke-[--text-normal]'}"
     viewBox="30 29 300 300"
 >
     <defs>
