@@ -1,8 +1,8 @@
 <script lang="ts">
-    let index = 0;
-    setInterval(() => {
-        index = (index + 1) % 4;
-    }, 300);
+let index = $state(0);
+setInterval(() => {
+	index = (index + 1) % 4;
+}, 300);
 </script>
 
 {#each ['', '.', '..', '...'] as sequence, i}
