@@ -1,23 +1,12 @@
-import { init, addMessages } from 'svelte-i18n';
+import { init, addMessages } from "svelte-i18n";
 
-import en from './en.json';
-import es from './es.json';
-import tr from './tr.json';
-import fr from './fr.json';
-import zh from './zh.json';
-import ko from './ko.json';
-
+import en from "./en.json";
 // import de from './de.json';
 
-addMessages('en', en);
-addMessages('es', es);
-addMessages('tr', tr);
-addMessages('fr', fr);
-addMessages('zh', zh);
-addMessages('ko', ko);
+addMessages("en", en);
 // addMessages('de', de);
 
 init({
-    fallbackLocale: 'en',
-    initialLocale: window.localStorage.getItem('language'),
+	fallbackLocale: "en",
+	initialLocale: window.localStorage.getItem("language"),
 });
