@@ -1,277 +1,322 @@
-import { around } from 'monkey-around';
-import { Notice, Plugin, TFile, WorkspaceLeaf, WorkspaceSidedock, normalizePath, type ViewState } from 'obsidian';
-import { LogLvl, Prompts, type Language, type OllamaEmbedModel, type OllamaGenModel, type OpenAIEmbedModel, type OpenAIGenModel } from 'papa-ts';
-import { get } from 'svelte/store';
-import { _ } from 'svelte-i18n';
+import { after, around } from "monkey-around";
+import {
+  Modal,
+  Notice,
+  Plugin,
+  TAbstractFile,
+  TFile,
+  WorkspaceLeaf,
+  WorkspaceSidedock,
+  normalizePath,
+  type ViewState,
+} from "obsidian";
+import {
+  AnthropicProvider,
+  CustomOpenAIProvider,
+  LogLvl,
+  OllamaProvider,
+  OpenAIProvider,
+  type Language,
+  type Providers,
+  type RegisteredProvider,
+  type RegisteredEmbedProvider,
+  type RegisteredGenProvider,
+  Papa,
+} from "papa-ts";
+import { get } from "svelte/store";
+import { _ } from "svelte-i18n";
 
-import SmartSecondBrain from './SmartSecondBrain';
-import { ConfirmModal } from './components/Settings/ConfirmModal';
-import { PullModal } from './components/Modal/PullModal';
-import './lang/i18n';
-import Log from './logging';
-import { chatHistory, data, isChatInSidebar, plugin } from './store';
-import './styles.css';
-import { ChatView, VIEW_TYPE_CHAT } from './views/Chat';
-import { SetupView, VIEW_TYPE_SETUP } from './views/Onboarding';
-import SettingsTab from './views/Settings';
-import { RemoveModal } from './components/Modal/RemoveModal';
+import { ConfirmModal } from "./components/Settings/ConfirmModal";
+import "./lang/i18n";
+import Log from "./logging";
+import "./styles.css";
+import SettingsTab from "./views/Settings/Settings";
+import type { ProviderConfigs } from "./types/providers";
+import {
+  createData,
+  getData,
+  PluginDataStore,
+} from "./stores/dataStore.svelte";
+import { chatLayout, setPlugin } from "./stores/state.svelte";
+import { QueryClient } from "@tanstack/svelte-query";
+import { ChatView, VIEW_TYPE_CHAT } from "./views/Chat/Chat";
+import { ChatDB } from "./db/chatDbSchema";
+import { type ChatModel, createMessenger } from "./stores/chatStore.svelte";
+import type { UUIDv7 } from "./utils/uuid7Validator";
 
 export interface PluginData {
-    isChatComfy: boolean;
-    initialAssistantMessageContent: string;
-    isUsingRag: boolean;
-    assistantLanguage: Language;
-    isIncognitoMode: boolean;
-    ollamaGenModel: OllamaGenModel;
-    ollamaEmbedModel: OllamaEmbedModel;
-    openAIGenModel: OpenAIGenModel;
-    openAIEmbedModel: OpenAIEmbedModel;
-    targetFolder: string;
-    excludeFF: Array<string>;
-    defaultChatName: string;
-    debugginLangchainKey: string;
-    isQuickSettingsOpen: boolean;
-    isVerbose: boolean;
-    isOnboarded: boolean;
-    hideIncognitoWarning: boolean;
-    isAutostart: boolean;
+  providerConfig: ProviderConfigs;
+  initialAssistantMessageContent: string;
+  isUsingRag: boolean;
+  isGeneratingChatTitle: boolean;
+  defaultChatModel: ChatModel | null;
+  retrieveTopK: number;
+  assistantLanguage: Language;
+  excludeFF: Array<string>;
+  includeFF: Array<string>;
+  isExcluding: boolean;
+  defaultChatName: string;
+  debuggingLangchainKey: string;
+  isQuickSettingsOpen: boolean;
+  isVerbose: boolean;
+  hideIncognitoWarning: boolean;
+  isAutostart: boolean;
+  lastActiveChatId: UUIDv7 | null;
 }
 
-export type PluginDataKey = keyof PluginData;
-
-export const DEFAULT_SETTINGS: Partial<PluginData> = {
-    isChatComfy: true,
-    isUsingRag: true,
-    assistantLanguage: (window.localStorage.getItem('language') as Language) || 'en',
-    initialAssistantMessageContent:
-        Prompts[(window.localStorage.getItem('language') as Language) || 'en']?.initialAssistantMessage || Prompts.en.initialAssistantMessage,
-    isIncognitoMode: true,
-    ollamaGenModel: { model: 'llama2', baseUrl: 'http://localhost:11434', temperature: 0.5 },
-    ollamaEmbedModel: {
-        model: 'nomic-embed-text',
-        baseUrl: 'http://localhost:11434',
-        similarityThreshold: 0.75,
-        k: 100,
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      refetchOnWindowFocus: false,
+      refetchOnReconnect: false,
+      retry: 0,
     },
-    openAIGenModel: {
-        model: 'gpt-3.5-turbo',
-        openAIApiKey: '',
-        temperature: 0.5,
-    },
-    openAIEmbedModel: {
-        model: 'text-embedding-ada-002',
-        openAIApiKey: '',
-        similarityThreshold: 0.75,
-        k: 100,
-    },
-    targetFolder: 'Chats',
-    defaultChatName: 'New Chat',
-    excludeFF: ['Chats', '*.excalidraw.md'],
-    isQuickSettingsOpen: true,
-    isVerbose: false,
-    isOnboarded: false,
-    hideIncognitoWarning: false,
-    isAutostart: false,
-};
+  },
+});
 
 export default class SecondBrainPlugin extends Plugin {
-    setupView: SetupView;
-    chatView: ChatView;
-    s2b: SmartSecondBrain;
-    leaf: WorkspaceLeaf;
-    private isChatAcivatedFromRibbon = false; // workaround for a bug where the chat view is activated twice through monkey patching
-    private autoSaveTimer: number;
+  papa!: Papa;
+  queryClient = queryClient;
+  pluginData!: PluginDataStore;
+  private isChatAcivatedFromRibbon = false; // workaround for a bug where the chat view is activated twice through monkey patching
+  private autoSaveTimer!: number;
+  private chatCacheDb!: ChatDB;
 
-    async loadSettings() {
-        data.set(Object.assign({}, DEFAULT_SETTINGS, await this.loadData()));
+  async onload() {
+    setPlugin(this);
+    const VIEW_TYPE_CHAT = "chat-view";
+    this.pluginData = await createData(this);
+    this.chatCacheDb = new ChatDB();
+    const messenger = createMessenger(this.chatCacheDb);
+
+    this.papa = new Papa({
+      debugging: {
+        langsmithApiKey: this.pluginData.debuggingLangchainKey,
+        logLvl: 1,
+      },
+    });
+
+    //Todo setup method with preamptife fetching etcs
+    this.pluginData.getConfiguredProviders().forEach((provider) => {
+      this.papa.providerRegistry
+        .getProvider(provider)
+        .setup(this.pluginData.getProviderAuthParams(provider));
+    });
+
+    const { isVerbose, isAutostart } = this.pluginData;
+
+    if (this.manifest.dir === undefined) {
+      this.unload();
+      throw Error("Cannot localize plugin directory.");
     }
 
-    async saveSettings() {
-        await this.saveData(get(data));
+    // this.s2b = new SmartSecondBrain(this, this.manifest.dir);
+    // Log.setLogLevel(isVerbose ? LogLvl.DEBUG : LogLvl.DISABLED);
+
+    // this.app.workspace.onLayoutReady(() => {
+    //   if (isAutostart) this.s2b.init();
+    // });
+
+    // this.registerEvent(
+    // 	this.app.workspace.on("layout-change", () => {
+    // 		if (!this.leaf) {
+    // 			const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT);
+    // 			if (!leaves.length) return;
+    // 			this.leaf = leaves[0];
+    // 		}
+    // 		chatLayout.isSidebar = [this.app.workspace.leftSplit, this.app.workspace.rightSplit].includes(
+    // 			this.leaf.getRoot() as WorkspaceSidedock,
+    // 		);
+    // 	}),
+    // );
+    // this.registerEvent(
+    //   this.app.metadataCache.on("changed", async (file: TFile) =>
+    //     this.s2b.onFileChange(file),
+    //   ),
+    // );
+    // this.registerEvent(
+    //   this.app.vault.on("delete", async (file: TAbstractFile) => {
+    //     // guard: only handle real files
+    //     if (!(file instanceof TFile)) return;
+    //     await this.s2b.onFileDelete(file);
+    //   }),
+    // );
+    // this.registerEvent(
+    //   this.app.vault.on(
+    //     "rename",
+    //     async (file: TAbstractFile, oldPath: string) => {
+    //       // guard: only handle real files
+    //       if (!(file instanceof TFile)) return;
+    //       await this.s2b.onFileRename(file, oldPath);
+    //     },
+    //   ),
+    // );
+    // // periodically or on unfocus save vector store data to disk
+    // window.addEventListener("blur", () => this.s2b.saveVectorStoreData());
+
+    // const resetAutoSaveTimer = () => {
+    //   window.clearTimeout(this.autoSaveTimer);
+    //   this.autoSaveTimer = window.setTimeout(
+    //     () => this.s2b.saveVectorStoreData(),
+    //     30 * 1000,
+    //   );
+    // };
+    // listen for any event to reset the timer
+    // window.addEventListener("mousemove", () => resetAutoSaveTimer());
+    // window.addEventListener("mousedown", () => resetAutoSaveTimer());
+    // window.addEventListener("keypress", () => resetAutoSaveTimer());
+    // window.addEventListener("scroll", () => resetAutoSaveTimer());
+
+    this.registerView(VIEW_TYPE_CHAT, (leaf) => new ChatView(this, leaf));
+
+    this.addRibbonIcon("message-square", "ribbon.chat", () =>
+      this.activateView(),
+    );
+
+    // this.addCommand({
+    //   id: "open-chat",
+    //   name: t("cmd.chat"),
+    //   icon: "message-square",
+    //   callback: () => this.activateView(),
+    // });
+    // this.addCommand({
+    //   id: "pull-model",
+    //   name: t("cmd.pull_model"),
+    //   icon: "arrow-down-to-line",
+    //   callback: () => new PullModal(this.app).open(),
+    // });
+    // this.addCommand({
+    //   id: "remove-model",
+    //   name: t("cmd.remove_model"),
+    //   icon: "trash",
+    //   callback: () => new RemoveModal(this.app).open(),
+    // });
+
+    this.addSettingTab(new SettingsTab(this));
+
+    //this.registerMonkeyPatches();
+  }
+
+  async onunload() {
+    Log.info("Unloading plugin");
+  }
+
+  async activateView() {
+    const { targetFolder, defaultChatName, initialAssistantMessageContent } =
+      this.pluginData;
+    // If no file is provided, open the default chat
+
+    const chatDirExists: boolean = !!this.app.vault.getFolderByPath(
+      normalizePath(targetFolder),
+    );
+
+    if (!chatDirExists) {
+      await this.app.vault.createFolder(normalizePath(targetFolder));
     }
 
-    async onload() {
-        plugin.set(this);
-        const t = get(_);
-        await this.loadSettings();
-        this.s2b = new SmartSecondBrain(this.app, this.manifest.dir);
-        Log.setLogLevel(get(data).isVerbose ? LogLvl.DEBUG : LogLvl.DISABLED);
+    const filePath = normalizePath(`${targetFolder}/${defaultChatName}.md`);
 
-        this.app.workspace.onLayoutReady(() => {
-            const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT) || this.app.workspace.getLeavesOfType(VIEW_TYPE_SETUP);
-            if (leaves.length) this.leaf = leaves[0];
-            if (get(data).isOnboarded && get(data).isAutostart) this.s2b.init();
-        });
-        this.registerEvent(
-            this.app.workspace.on('layout-change', () => {
-                if (!this.leaf) {
-                    const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT) || this.app.workspace.getLeavesOfType(VIEW_TYPE_SETUP);
-                    if (!leaves.length) return;
-                    this.leaf = leaves[0];
-                }
-                const isInSidebar = [this.app.workspace.leftSplit, this.app.workspace.rightSplit].includes(this.leaf.getRoot() as WorkspaceSidedock);
-                isChatInSidebar.set(isInSidebar);
-            })
-        );
-        this.registerEvent(this.app.metadataCache.on('changed', async (file: TFile) => this.s2b.onFileChange(file)));
-        this.registerEvent(this.app.vault.on('delete', async (file: TFile) => this.s2b.onFileDelete(file)));
-        this.registerEvent(this.app.vault.on('rename', async (file: TFile, oldPath: string) => this.s2b.onFileRename(file, oldPath)));
+    let file = this.app.vault.getFileByPath(filePath);
+    if (!file) {
+      file = await this.app.vault.create(
+        filePath,
+        "Assistant\n" + initialAssistantMessageContent + "\n- - - - -",
+      );
+    }
 
-        // periodically or on unfocus save vector store data to disk
-        window.addEventListener('blur', () => this.s2b.saveVectorStoreData());
+    const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT);
+    const isNewLeaf: boolean = !leaves.length;
+    const leaf = isNewLeaf ? this.app.workspace.getRightLeaf(false) : leaves[0];
+    if (isNewLeaf)
+      await leaf?.setViewState({ type: VIEW_TYPE_CHAT, active: true });
+    this.app.workspace.revealLeaf(leaf);
+  }
 
-        const resetAutoSaveTimer = () => {
-            window.clearTimeout(this.autoSaveTimer);
-            this.autoSaveTimer = window.setTimeout(() => this.s2b.saveVectorStoreData(), 30 * 1000);
-        };
-        // listen for any event to reset the timer
-        window.addEventListener('mousemove', () => resetAutoSaveTimer());
-        window.addEventListener('mousedown', () => resetAutoSaveTimer());
-        window.addEventListener('keypress', () => resetAutoSaveTimer());
-        window.addEventListener('scroll', () => resetAutoSaveTimer());
+  // async saveChat() {
+  // 	const { targetFolder } = this.pluginData;
+  // 	let fileName = await this.s2b.createFilenameForChat();
+  // 	let normalizedFilePath = normalizePath(targetFolder + "/" + fileName + ".md");
+  // 	while (await this.app.vault.adapter.exists(normalizedFilePath)) {
+  // 		//Checks if already existing file has a number at the end
+  // 		const regex = /\((\d+)\)$/;
+  // 		const match = fileName.match(regex);
+  // 		if (match) {
+  // 			fileName = fileName.slice(0, -3) + "(" + (parseInt(match[1], 10) + 1) + ")";
+  // 		} else {
+  // 			fileName = fileName + " (1)";
+  // 		}
+  // 		normalizedFilePath = normalizePath(d.targetFolder + "/" + fileName + ".md");
+  // 	}
+  // 	// const newChatFile = await this.app.vault.copy(this.chatView.file, normalizedFilePath);
+  // 	chatHistory.reset;
+  // 	await this.activateView(newChatFile);
+  // }
 
-        if (!this.s2b) {
-            this.app.workspace.detachLeavesOfType(VIEW_TYPE_CHAT);
+  async clearPluginData() {
+    const t = get(_);
+    new ConfirmModal(
+      this.app,
+      t("settings.clear_modal.title"),
+      t("settings.clear_modal.description"),
+      async (result) => {
+        if (result === "Yes") {
+          const { deleteData } = this.pluginData;
+          deleteData();
+          const files = (
+            await this.app.vault.adapter.list(
+              normalizePath(this.manifest.dir + "/vectorstores"),
+            )
+          ).files;
+          for (const file of files) await this.app.vault.adapter.remove(file);
+          new Notice(t("notice.plugin_data_cleared"), 4000);
+          this.pluginData = await createData(this);
+          await this.activateView();
         }
+      },
+      "",
+    ).activate();
+  }
 
-        this.registerView(VIEW_TYPE_SETUP, (leaf) => {
-            this.setupView = new SetupView(leaf);
-            return this.setupView;
-        });
+  // registerMonkeyPatches() {
+  //   const { targetFolder } = this.pluginData;
+  //   // eslint-disable-next-line @typescript-eslint/no-this-alias
+  //   const self = this;
+  //   // Monkey patch WorkspaceLeaf to open Chats with ChatView by default
+  //   this.register(
+  //     around(WorkspaceLeaf.prototype, {
+  //       setViewState(next) {
+  //         // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  //         return function (state: ViewState, ...rest: any[]) {
+  //           if (
+  //             // If we have a markdown file
+  //             state.type === "markdown" &&
+  //             state.state?.file
+  //           ) {
+  //             // Then check for the chat frontMatterKey
+  //             // const cache = self.app.metadataCache.getCache(state.state.file);
 
-        this.registerView(VIEW_TYPE_CHAT, (leaf) => {
-            this.chatView = new ChatView(leaf);
-            return this.chatView;
-        });
+  //             // if (cache?.frontmatter && cache.frontmatter['second-brain-chat']) {
+  //             if (state.state.file.startsWith(targetFolder)) {
+  //               // If we have it, force the view type to be chat
+  //               const newState = {
+  //                 ...state,
+  //                 type: VIEW_TYPE_CHAT,
+  //               };
+  //               if (!self.isChatAcivatedFromRibbon) {
+  //                 const leaves =
+  //                   self.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT);
+  //                 self.leaf = leaves.length
+  //                   ? leaves[0]
+  //                   : self.app.workspace.getRightLeaf(false);
+  //               }
+  //               return next.apply(self.leaf, [newState, ...rest]);
+  //             }
+  //           }
 
-        this.addRibbonIcon('message-square', t('ribbon.chat'), () => this.activateView());
-
-        this.addCommand({ id: 'open-chat', name: t('cmd.chat'), icon: 'message-square', callback: () => this.activateView() });
-
-        this.addCommand({ id: 'pull-model', name: t('cmd.pull_model'), icon: 'arrow-down-to-line', callback: () => new PullModal(this.app).open() });
-
-        this.addCommand({ id: 'remove-model', name: t('cmd.remove_model'), icon: 'trash', callback: () => new RemoveModal(this.app).open() });
-
-        this.addSettingTab(new SettingsTab(this.app, this));
-
-        this.registerMonkeyPatches();
-    }
-
-    async onunload() {
-        Log.info('Unloading plugin');
-    }
-
-    async activateView(file?: TFile) {
-        const d = get(data);
-        if (!d.isOnboarded) {
-            this.app.workspace.detachLeavesOfType(VIEW_TYPE_CHAT);
-            const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_SETUP);
-            this.leaf = leaves.length ? leaves[0] : this.app.workspace.getRightLeaf(false);
-            await this.leaf.setViewState({ type: VIEW_TYPE_SETUP, active: true });
-        } else {
-            this.app.workspace.detachLeavesOfType(VIEW_TYPE_SETUP);
-            if (!file) {
-                // If no file is provided, open the default chat
-                const chatDirExists = await this.app.vault.adapter.exists(normalizePath(d.targetFolder));
-                if (!chatDirExists) {
-                    await this.app.vault.createFolder(normalizePath(d.targetFolder));
-                }
-                const defaultChatExists = await this.app.vault.adapter.exists(normalizePath(d.targetFolder + '/' + d.defaultChatName + '.md'));
-                file = defaultChatExists
-                    ? this.app.metadataCache.getFirstLinkpathDest(d.targetFolder + '/' + d.defaultChatName + '.md', '')
-                    : await this.app.vault.create(
-                          normalizePath(d.targetFolder + '/' + d.defaultChatName + '.md'),
-                          'Assistant\n' + d.initialAssistantMessageContent + '\n- - - - -'
-                      );
-            }
-            const leaves = this.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT);
-            this.leaf = leaves.length ? leaves[0] : this.app.workspace.getRightLeaf(false);
-            this.isChatAcivatedFromRibbon = true;
-            await this.leaf.openFile(file, { active: true });
-            await this.leaf.setViewState({
-                type: VIEW_TYPE_CHAT,
-                state: { file: file.path },
-            });
-        }
-        this.app.workspace.revealLeaf(this.leaf);
-    }
-
-    async saveChat() {
-        const d = get(data);
-        let fileName = await this.s2b.createFilenameForChat();
-        let normalizedFilePath = normalizePath(d.targetFolder + '/' + fileName + '.md');
-        while (await this.app.vault.adapter.exists(normalizedFilePath)) {
-            //Checks if already existing file has a number at the end
-            const regex = /\((\d+)\)$/;
-            const match = fileName.match(regex);
-            if (match) {
-                fileName = fileName.slice(0, -3) + '(' + (parseInt(match[1], 10) + 1) + ')';
-            } else {
-                fileName = fileName + ' (1)';
-            }
-            normalizedFilePath = normalizePath(d.targetFolder + '/' + fileName + '.md');
-        }
-        const newChatFile = await this.app.vault.copy(this.chatView.file, normalizedFilePath);
-        chatHistory.reset;
-        await this.activateView(newChatFile);
-    }
-
-    async clearPluginData() {
-        const t = get(_);
-        new ConfirmModal(
-            get(plugin).app,
-            t('settings.clear_modal.title'),
-            t('settings.clear_modal.description'),
-            async (result) => {
-                if (result === 'Yes') {
-                    await this.saveData({});
-                    const files = (await this.app.vault.adapter.list(normalizePath(this.manifest.dir + '/vectorstores'))).files;
-                    for (const file of files) await this.app.vault.adapter.remove(file);
-                    new Notice(t('notice.plugin_data_cleared'), 4000);
-                    await this.loadSettings();
-                    await this.activateView();
-                }
-            },
-            ''
-        ).activate();
-    }
-
-    registerMonkeyPatches() {
-        // eslint-disable-next-line @typescript-eslint/no-this-alias
-        const self = this;
-        // Monkey patch WorkspaceLeaf to open Chats with ChatView by default
-        this.register(
-            around(WorkspaceLeaf.prototype, {
-                setViewState(next) {
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    return function (state: ViewState, ...rest: any[]) {
-                        if (
-                            // If we have a markdown file
-                            state.type === 'markdown' &&
-                            state.state?.file
-                        ) {
-                            // Then check for the chat frontMatterKey
-                            // const cache = self.app.metadataCache.getCache(state.state.file);
-
-                            // if (cache?.frontmatter && cache.frontmatter['second-brain-chat']) {
-                            if (state.state.file.startsWith(get(data).targetFolder)) {
-                                // If we have it, force the view type to be chat
-                                const newState = {
-                                    ...state,
-                                    type: VIEW_TYPE_CHAT,
-                                };
-                                if (!self.isChatAcivatedFromRibbon) {
-                                    const leaves = self.app.workspace.getLeavesOfType(VIEW_TYPE_CHAT);
-                                    self.leaf = leaves.length ? leaves[0] : self.app.workspace.getRightLeaf(false);
-                                }
-                                return next.apply(self.leaf, [newState, ...rest]);
-                            }
-                        }
-
-                        return next.apply(this, [state, ...rest]);
-                    };
-                },
-            })
-        );
-    }
+  //           return next.apply(this, [state, ...rest]);
+  //         };
+  //       },
+  //     }),
+  //   );
+  // }
 }
